requirements.txt
# Core frameworks
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# PDF processing
pdfplumber>=0.9.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0

# Date parsing
python-dateutil>=2.8.0

# Machine Learning & Embeddings
langchain>=0.1.0
langchain-core>=0.1.0
langchain-huggingface>=0.0.1
langchain-pinecone>=0.1.0
langchain-google-genai>=1.0.0

# Vector Database
pinecone-client>=3.0.0

# Environment variables
python-dotenv>=1.0.0

# Additional dependencies for Hugging Face transformers
transformers>=4.0.0
sentence-transformers>=2.0.0
torch>=1.9.0
